Revert getTransaction status check.
The JSON RPC spec returns a status of 0x1 or 0x0 if the transaction
succeeded or failed. Web3 v1.0.0-beta.33 preserves this hex value so it
must be parsed to figure whether the operation succeeded or not.

This changes in the following versions of web3 where the status is
casted to a boolean value:

https://github.com/ethereum/web3.js/blob/v1.0.0-beta.34/packages/web3-core-helpers/src/formatters.js#L219-L221