Pause the transaction queue if there is not enough balance to operate.
I added a balance check (for both WIB and ETH) before sending any transaction
to the blockchain to pause the transaction queue. As a result of this,
transaction jobs get queued up and will eventually be processed once the
queue is resumed.

The pausing of the transaction queue occurs while processing a job. This job
will be run until it's finished but no other jobs will be processed.

To resume the transaction queue a task must be run:

```
$ npm run task -- txqueue:resume
```

Also, there are other tasks that can be run:

```
$ npm run task -- txqueue:pause
$ npm run task -- txqueue:count
```